refactor LSF class to be compatible with batch
